pyproject.toml
[build-system]
requires = ["setuptools>=69", "wheel"]
build-backend = "setuptools.build_meta"

[project]
name = "mcpwn-red"
version = "0.1.0"
description = "Adversarial safety harness for the MCPwn AI pentesting execution engine"
readme = "README.md"
requires-python = ">=3.11"
license = { text = "GPL-3.0-only" }
authors = [
  { name = "OpenAI Codex" },
]
classifiers = [
  "Development Status :: 4 - Beta",
  "Environment :: Console",
  "Intended Audience :: Information Technology",
  "Intended Audience :: System Administrators",
  "License :: OSI Approved :: GNU General Public License v3 (GPLv3)",
  "Operating System :: POSIX :: Linux",
  "Programming Language :: Python :: 3",
  "Programming Language :: Python :: 3.11",
  "Programming Language :: Python :: 3.12",
  "Topic :: Security",
  "Topic :: System :: Monitoring",
]
dependencies = [
  "anyio>=4.0",
  "click>=8.1",
  "httpx>=0.27",
  "jinja2>=3.1",
  "mcp>=1.0",
  "pydantic>=2.0",
  "pyyaml>=6.0",
  "rich>=13.0",
]

[project.optional-dependencies]
dev = [
  "mypy>=1.8",
  "pytest>=7.0",
  "pytest-asyncio>=0.23",
  "respx>=0.22",
  "ruff>=0.5",
]

[project.scripts]
mcpwn-red = "mcpwn_red.cli:main"

[tool.setuptools]
include-package-data = true

[tool.setuptools.packages.find]
include = ["mcpwn_red*"]

[tool.pytest.ini_options]
asyncio_mode = "auto"
testpaths = ["tests"]

[tool.ruff]
line-length = 100
target-version = "py311"

[tool.ruff.lint]
select = [
  "E",
  "F",
  "I",
  "N",
  "UP",
  "B",
  "S",
  "A",
  "C4",
  "DTZ",
  "T20",
]
ignore = [
  "S101",
  "E501",
  "I001",
  "UP017",
  "UP035",
  "UP037",
  "UP041",
]

[tool.mypy]
python_version = "3.11"
warn_unused_configs = true
warn_redundant_casts = true
warn_unused_ignores = true
strict_equality = true
no_implicit_optional = true
disallow_untyped_defs = true
disallow_incomplete_defs = true
check_untyped_defs = true
plugins = ["pydantic.mypy"]
